code cleanup

+++ src/options.cpp
@@ -135,11 +135,11 @@ TimeControl Options::parseTc(const std::string tcString)
     if (has_inc)
     {
         auto moves = splitString(remainingStringVector, '+');
-        tc.increment = std::stof(moves[1].c_str()) * 1000;
+        tc.increment = static_cast<int64_t>(std::stof(moves[1].c_str())) * 1000;
         remainingStringVector = moves[0];
     }
 
-    tc.time = std::stof(remainingStringVector.c_str()) * 1000;
+    tc.time = static_cast<int64_t>(std::stof(remainingStringVector.c_str())) * 1000;
 
     return tc;
 }
